Add Animation to objects
-Kinda buggy when using bot if botDelay differs from animationDelay
-TODO: change setBotDelay to set animationDelay also.
-TODO: Add attack effect
-TODO: Animate first player then enemies

+++ Meta_Engine/include/Player.h
@@ -14,7 +14,7 @@ class Player : public Entity
         virtual ~Player();
 
         void draw();
-        void update(unsigned int dt = 0);
+        void update(float dt = 0);
         void events(sf::Event& event);
         void startController();
         void resetMoved();
