+++ Meta_Engine/include/State.h
@@ -20,7 +20,7 @@ class State
     public:
         virtual void load(int stack = 0) = 0;
         virtual int unload() = 0;
-        virtual eStateType update(unsigned int dt) = 0;
+        virtual eStateType update(float dt) = 0;
         virtual void render() = 0;
         virtual void events(sf::Event& event) = 0;
         virtual ~State(){};
