+++ Meta_Engine/include/StateGame.h
@@ -8,7 +8,7 @@ class StateGame : public State
         virtual ~StateGame();
         void load(int stack = 0);
         int unload();
-        eStateType update(unsigned int dt);
+        eStateType update(float dt);
         void events(sf::Event& event);
         void render();
     protected:
